Reorganize pmid reading
