+++ indra/tools/reading/pmid_reading/run_aws.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, print_function, unicode_literals
 from builtins import dict, str
-from indra.tools.reading.read_pmids import READER_DICT
+from . import READER_DICT
 
 DOC = \
 """
